+++ test/support/test_case.ex
@@ -1,0 +1,9 @@
+defmodule Sci.TestCase do
+  use ExUnit.CaseTemplate
+
+  using do
+    quote do
+      import ComplexNumSigil
+    end
+  end
+end